requirements.txt
asgiref
boto3
crispy-tailwind
Django
django-crispy-forms
django-environ
django-filter
django-phonenumber-field
django-storages
djangorestframework
gunicorn
importlib-metadata
Markdown
phonenumberslite
psycopg2
pytz
sqlparse
zipp
